requirements.txt
pdfplumber==0.11.4
anthropic>=0.25.0
openai==1.56.1
python-dotenv==1.0.1
pytesseract==0.3.13
pdf2image==1.17.0
pypdfium2==4.30.0
Pillow==10.4.0
streamlit==1.40.0